+++ tests/Sequence/DeferTest.php
@@ -314,6 +314,23 @@ class DeferTest extends TestCase
         );
     }
 
+    public function testLastWhenNull()
+    {
+        $sequence = new Defer((static function() {
+            yield 1;
+            yield 2;
+            yield 3;
+            yield null;
+        })());
+
+        $this->assertNull(
+            $sequence->last()->match(
+                static fn($value) => $value,
+                static fn() => false,
+            ),
+        );
+    }
+
     public function testContains()
     {
         $sequence = new Defer((static function() {